feat: optimize privateKey usage
