+++ src/config/index.ts
@@ -1,19 +1,16 @@
-export interface IConfig {
-  address: string
-  privateKey: string
-  providerUrl: string
-}
+import { IConfig, SignRawTransactionFnParams } from '../global'
 
 const config = {
   address: '',
-  privateKey: '',
   providerUrl: '',
+  personalSignFn(_privateKey: string) { return '' },
+  signRawTransactionFn(_params: SignRawTransactionFnParams) { return '' },
 }
 
-export const setConfig = (params: IConfig) => {
+export const setConfig = (params: IConfig): IConfig => {
   return Object.assign(config, params)
 }
 
-export const getConfig = () => {
+export const getConfig = (): IConfig => {
   return config
 }
